Ground & Land
Had to change the layer system, had to add another variable for ground layer as vector<Land>
Had to do this bc otherwise Land's getX and getY can't be used due to the vector being GameObject*

Because of this, Land had to be changed as well. It's no longer a GameObject, as GOs need Layer as a parent
(which means vector<Land> isn't possible). It now has a Ground class as a parent.

Ground class was introduced as Land's parent class.

LayerSystem now has a 3d vector of GOs (a vector of Layers) and a 2d vector of Land (Ground)